feat: divide functions into files
