+++ go.mod
@@ -8,4 +8,4 @@ require (
 	github.com/toadharvard/stopwords-iso v0.1.5
 )
 
-require gopkg.in/yaml.v3 v3.0.1 // indirect
+require gopkg.in/yaml.v3 v3.0.1
